Added an interactive procedure, with dialog package. It install 'dialog' package if not found and then start asking questions.
The interactive procedure will happen if the `-N` option isnt provided.
If provided, the user need to specify them with command line options.

When in interactive mode, a main menu will allow the user either install
or uninstall floresta from system. If user choose to install, it will
show a menu that will help to add (or not add):

* the type of network (bitcoin, sigtest, testnet, regtest);
* a node to connect to (ip:port);
* proxy (http, https, socks5 or simple ip/port format);
* enable cfilters;
* assume-valid (with a blockheight to choose);
* assume-utreexo
* ssl (with generation of a key and certificate paths and electrum
ip/port);
* addition of xpubs;
* addition of descriptors; and
* review message box;

After review, the user can confirm (or not confirm) the installation
procedure.